requirements.txt
django
djangorestframework
psutil
prometheus-client
requests
dotenv
gunicorn